+++ sdk/security/keyvault/azsecrets/CHANGELOG.md
@@ -1,8 +1,9 @@
 # Release History
 
 ## 1.5.1-beta.1 (Unreleased)
-
 ### Features Added
+
+* Added support for API version `2026-03-01-preview` (now the default).
 
 ### Breaking Changes